+++ types.d.ts
@@ -347,6 +347,8 @@ type WatchFiles = {
   paths?: Array<string>;
   files?: Array<RegExp>;
   ignore?: Array<RegExp>;
+  includes?: Array<RegExp | string>;
+  excludes?: Array<RegExp | string>;
 };
 
 // Private types: used in source code only